// backend/src/modules/mpesa/types.ts
export interface MpesaModuleOptions {
    consumerKey: string
    consumerSecret: string
    shortcode: string
    passkey: string
    initiatorName: string
    securityCredential: string
    callbackBaseUrl: string
    environment: "sandbox" | "production"
    redisUrl: string
    slackWebhookUrl?: string
    financeAdminEmail?: string
}